Mark adjacent grids that must be generated and then force generation if necessary
The fact that a hierarchy is empty doesn't necessarily means that none of its children 2 or more levels down will also be empty. It may be the case that the resolution of the hierarchy grids is too low. If the generation of a hierarchy node has been stopped because it's empty but then later a child needs to be generated, the path from where generation stopped to where it's needed, is marked. Any marked Hierarchy node will allow continued generation no matter if it's empty or not.

Each node in the octree has 8 children representing a 2x2x2 of subgrids i.e. a 3d array. 1 bit is needed to represent 1 index in the 3d array because it's only 2 long. Essentially for each level in the tree, 3 bits are necessary to store its path from root to node. GridPos is a new class that stores this information in 3 variables X, Y and Z. each lsb of the variables represents the three indices of the parent node. Going down the tree will shift the variables 1 to the left and the indices are again written to lsb. Moving on the X axis is done by incrementing or decrementing the X variable. Incrementing the X path 101 will give 110 which is the correct path to the positive x adjacent grid.
GridPos.TryMove checks if moving the GridPos in a direction will move it outside the bounds of the tree and of so, it will return false.